pyproject.toml
[tool.poetry]
name = "data-enrichment-agents"
version = "0.1.0"
description = "Data-enrichment tool for leads list using AI Research Agents"
authors = ["Isaac-Porat"]
license = "GPL-3.0 license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
llama-index = "^0.11.18"
langchain-community = "^0.3.2"
llama-index-readers-web = "^0.2.4"
firecrawl-py = "^1.3.1"
anthropic = "^0.36.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
